pyproject.toml
[project]
name = "anime-suggestions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.0",
    "google-genai>=1.61.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "slowapi>=0.1.9",
    "streamlit>=1.53.1",
]
